requirements.txt
# requirements.txt
Flask
requests
pytest
lark-oapi
python-jenkins
openai
dashscope
mem0ai
python-dotenv